Move concurrency specific functionality into concurrency_aware_ait
Although the concurrent fix-point and merge_shared could be in
ai_baset, there is relatively little point.  Both are specific to
the way that concurrency_aware_ait works.  The analysis itself is
only sound if some relatively particular and quite limiting
conditions hold on the domain.  If we want to improve the way we
do abstract intepretation of concurrent programs I doubt we will
want to build on top of this.
